+++ DatadogRUM/Sources/RUMMonitor/Scopes/RUMSessionScope.swift
@@ -102,7 +102,7 @@ internal class RUMSessionScope: RUMScope, RUMContextProvider {
             didStartWithReplay: hasReplay
         )
 
-        if let viewScope = resumingViewScope, viewScope.identity.exists {
+        if let viewScope = resumingViewScope {
             viewScopes.append(
                 RUMViewScope(
                     isInitialView: false,
@@ -141,9 +141,6 @@ internal class RUMSessionScope: RUMScope, RUMContextProvider {
 
         // Transfer active Views by creating new `RUMViewScopes` for their identity objects:
         self.viewScopes = expiredSession.viewScopes.compactMap { expiredView in
-            guard expiredView.identity.exists else {
-                return nil // if the underlying identifiable (`UIVIewController`) no longer exists, skip transferring its scope
-            }
             return RUMViewScope(
                 isInitialView: false,
                 parent: self,
@@ -260,7 +257,7 @@ internal class RUMSessionScope: RUMScope, RUMContextProvider {
             isInitialView: true,
             parent: self,
             dependencies: dependencies,
-            identity: RUMOffViewEventsHandlingRule.Constants.applicationLaunchViewURL.asRUMViewIdentity(),
+            identity: ViewIdentifier(RUMOffViewEventsHandlingRule.Constants.applicationLaunchViewURL),
             path: RUMOffViewEventsHandlingRule.Constants.applicationLaunchViewURL,
             name: RUMOffViewEventsHandlingRule.Constants.applicationLaunchViewName,
             attributes: command.attributes,
@@ -305,7 +302,7 @@ internal class RUMSessionScope: RUMScope, RUMContextProvider {
                 isInitialView: isStartingInitialView,
                 parent: self,
                 dependencies: dependencies,
-                identity: RUMOffViewEventsHandlingRule.Constants.backgroundViewURL.asRUMViewIdentity(),
+                identity: ViewIdentifier(RUMOffViewEventsHandlingRule.Constants.backgroundViewURL),
                 path: RUMOffViewEventsHandlingRule.Constants.backgroundViewURL,
                 name: RUMOffViewEventsHandlingRule.Constants.backgroundViewName,
                 attributes: command.attributes,